LLMsPredict/requirements.txt
PyYAML
bitsandbytes
git+https://github.com/huggingface/transformers.git
git+https://github.com/huggingface/peft.git
git+https://github.com/huggingface/accelerate.git
torch
pandas
tqdm
loguru
huggingface-hub[cli]
scikit-learn
openpyxl
openai
sentencepiece
protobuf
